+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RenderTarget plugin (#1059, **@tomas7770**).
 - RenderPicker plugin (#1060, **@tomas7770**).
 - Task class, for use in asynchronous code (#1111, **@RiscadoA**).
+- HDR plugin (#1085, **@RiscadoA**).
 
 ### Changed